requirements.txt
flask>=2.0.0
prometheus-client>=0.12.0
python-dotenv>=0.19.0
pyyaml>=5.4.1
pytest>=7.0.0
gunicorn>=20.1.0
